requirements.txt
certifi==2020.12.5
chardet==4.0.0
idna==2.10
octorest==0.4
pkg-resources==0.0.0
python-dotenv==0.17.1
requests==2.25.1
six==1.16.0
urllib3==1.26.4
websocket-client==0.59.0
